+++ weblate/trans/views/files.py
@@ -149,16 +149,16 @@ def upload_translation(request, project, component, lang):
         author_email = form.cleaned_data["author_email"]
 
     # Check for overwriting
-    overwrite = False
+    conflicts = ""
     if request.user.has_perm("upload.overwrite", obj):
-        overwrite = form.cleaned_data["upload_overwrite"]
+        conflicts = form.cleaned_data["conflicts"]
 
     # Do actual import
     try:
         not_found, skipped, accepted, total = obj.merge_upload(
             request,
             request.FILES["file"],
-            overwrite,
+            conflicts,
             author_name,
             author_email,
             method=form.cleaned_data["method"],